% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_access.R, R/results_directions.R,
%   R/results_geocode.R
\name{access_result}
\alias{access_result}
\alias{direction_instructions}
\alias{direction_routes}
\alias{direction_legs}
\alias{direction_steps}
\alias{direction_points}
\alias{direction_polyline}
\alias{geocode_coordinates}
\alias{geocode_address}
\title{Access Result}
\usage{
access_result(res, result = c("instructions", "routes", "legs", "steps",
  "points", "polyline", "coordinates", "address"))

direction_instructions(res)

direction_routes(res)

direction_legs(res)

direction_steps(res)

direction_points(res)

direction_polyline(res)

geocode_coordinates(res)

geocode_address(res)
}
\arguments{
\item{res}{result from a Google API query}

\item{result}{the specific field of the result you want to access}
}
\description{
Methods for accessing specific elements of a Google API query.
}
\section{Functions}{
\itemize{
\item \code{direction_instructions}: the instructions from a directions query

\item \code{direction_routes}: the routes from a directions query

\item \code{direction_legs}: the legs from a directions query

\item \code{direction_steps}: the steps from a directions query

\item \code{direction_points}: the points from a directions query

\item \code{direction_polyline}: the encoded polyline from a direction query

\item \code{geocode_coordinates}: the coordinates from a geocode query

\item \code{geocode_address}: the formatted address from a geocode query
}}

\examples{
\dontrun{

apiKey <- "your_api_key"

## results returned as a list (simplify == TRUE)
lst <- google_directions(origin = c(-37.8179746, 144.9668636),
                        destination = c(-37.81659, 144.9841),
                        mode = "walking",
                        key = apiKey,
                        simplify = TRUE)

## results returned as raw JSON character vector
js <- google_directions(origin = c(-37.8179746, 144.9668636),
                         destination = c(-37.81659, 144.9841),
                        mode = "walking",
                         key = apiKey,
                         simplify = FALSE)

access_result(js, "polyline")

direction_polyline(js)

}
}
