% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_bicycling}
\alias{add_bicycling}
\title{Add bicycling}
\usage{
add_bicycling(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Adds bicycle route information to a googleway map object
}
\examples{
\dontrun{

google_map(key = "your_api_key") \%>\%
  add_bicycling()

}
}
