% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_circles}
\alias{add_circles}
\title{Add circle}
\usage{
add_circles(map, data, lat = NULL, lon = NULL, radius = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude.}

\item{lat}{string specifying the column of \code{data} containing the 'latitude' coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude' coordinates. If left NULL, a best-guess will be made}

\item{radius}{string specifying the column of \code{data} containing the 'radius' of each circle}
}
\description{
Add circles to a google map
}

