% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_drawing}
\alias{add_drawing}
\title{Add Drawing}
\usage{
add_drawing(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Adds drawing tools to the map
}
\examples{
\dontrun{

google_map(key = "your_api_key") \%>\%
  add_drawing()

}
}
