% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_kml}
\alias{add_kml}
\title{Add KML}
\usage{
add_kml(map, kml_data, layer_id = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{kml_data}{kml data layer}

\item{layer_id}{single value specifying an id for the layer.}
}
\description{
Adds a kml layer to a google map
}

