% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polygon}
\alias{add_polygon}
\title{Add polygon}
\usage{
add_polygon(map, data = get_map_data(map), polyline, stroke_colour = NULL,
  stroke_weight = NULL, stroke_opacity = NULL, fill_colour = NULL,
  fill_opacity = NULL, info_window = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{polyline}{string specifying the column containing the polyline}

\item{stroke_colour}{hex}

\item{stroke_weight}{num}

\item{stroke_opacity}{num}

\item{fill_colour}{hex}

\item{fill_opacity}{num}

\item{info_window}{string}
}
\description{
Add a polygon to a google map.
}

