% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{add_polyline}
\alias{add_polyline}
\title{Add polyline}
\usage{
add_polyline(map, data)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Add a polyline to a google map
}

