% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_google_polyline.R
\name{add_polyline_dev}
\alias{add_polyline_dev}
\title{Add polyline dev}
\usage{
add_polyline_dev(map, data = get_map_data(map), data_options = NULL,
  lat = NULL, lon = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{ordered data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{data_options}{data frame  -- including, ID, (must match an 'id' field in the \code{data}), and geodesic, strokeColor, strokeOpacity, strokeWeight}

\item{lat}{string specifying the column of \code{data} containing the 'latitude' coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude' coordinates. If left NULL, a best-guess will be made}
}
\description{
Add a polyline to a google map
}

