% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polylines}
\alias{add_polylines}
\title{Add polyline}
\usage{
add_polylines(map, data, polyline, geodesic = NULL, info_window = NULL,
  mouse_over = NULL, stroke_colour = NULL, stroke_weight = NULL,
  stroke_opacity = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{polyline}{string specifying the column of \code{data} containing the 'polyline'.}
}
\description{
Add a polyline to a google map
}

