% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_sf.R
\name{add_sf}
\alias{add_sf}
\title{Add sf}
\usage{
add_sf(map, data = get_map_data(map))
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the
latitude coordinates, and the other specifying the longitude. If Null, the
data passed into \code{google_map()} will be used.}
}
\description{
Adds an \code{sf} object to a google map
}
