% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_traffic}
\alias{add_traffic}
\title{Add Traffic}
\usage{
add_traffic(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Adds traffic information to a map
}

