% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_transit}
\alias{add_transit}
\title{Add transit}
\usage{
add_transit(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Adds public transport information to a googleway map object
}
\examples{
\dontrun{

map_key <- 'your_api_key'
google_map(key = map_key) \%>\%
  add_transit()

}
}
