% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{check_for_columns}
\alias{check_for_columns}
\title{Check for columns}
\usage{
check_for_columns(df, cols)
}
\arguments{
\item{df}{\code{data.frame}}

\item{cols}{string of columns}
}
\description{
Checks for valid columns
}

