% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{check_hex_colours}
\alias{check_hex_colours}
\title{Check hex colours}
\usage{
check_hex_colours(df, cols)
}
\arguments{
\item{df}{\code{data.frame}}

\item{cols}{string of columns to check}
}
\description{
Checks for valid hexadecimal value
}

