% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{check_opacities}
\alias{check_opacities}
\title{Check opacities}
\usage{
check_opacities(df, cols)
}
\arguments{
\item{df}{\code{data.frame}}

\item{cols}{string of columns to check}
}
\description{
Checks for valid opacity values
}

