% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_circle.R,
%   R/google_map_layer_drawing.R, R/google_map_layer_fusion.R,
%   R/google_map_layer_geojson.R, R/google_map_layer_heatmap.R,
%   R/google_map_layer_kml.R, R/google_map_layer_marker.R,
%   R/google_map_layer_polygon.R, R/google_map_layer_polyline.R,
%   R/google_map_layer_rectangle.R, R/google_map_layers.R
\name{clear_circles}
\alias{clear_circles}
\alias{clear_drawing}
\alias{remove_drawing}
\alias{clear_fusion}
\alias{clear_geojson}
\alias{clear_heatmap}
\alias{clear_kml}
\alias{clear}
\alias{clear_markers}
\alias{clear_polygons}
\alias{clear_polylines}
\alias{clear_rectangles}
\alias{clear_traffic}
\alias{clear_transit}
\alias{clear_bicycling}
\title{Remove drawing}
\usage{
clear_circles(map, layer_id = NULL)

clear_drawing(map)

remove_drawing(map)

clear_fusion(map, layer_id = NULL)

clear_geojson(map, layer_id = NULL)

clear_heatmap(map, layer_id = NULL)

clear_kml(map, layer_id = NULL)

clear_markers(map, layer_id = NULL)

clear_polygons(map, layer_id = NULL)

clear_polylines(map, layer_id = NULL)

clear_rectangles(map, layer_id = NULL)

clear_traffic(map)

clear_transit(map)

clear_bicycling(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{layer_id}{id value of the layer to be removed from the map}
}
\description{
Remove drawing

clears elements from a map
}
\section{Functions}{
\itemize{
\item \code{remove_drawing}: removes drawing controls from a map
}}

\note{
These operations are intended for use in conjunction with
\link{google_map_update} in an interactive shiny environment
}
