% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_bounds}
\alias{clear_bounds}
\title{Clear bounds}
\usage{
clear_bounds(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
A helper function to clear the javascript array of lat/lon bounds.
}
