% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_kml}
\alias{clear_kml}
\title{Clear kml}
\usage{
clear_kml(map, layer_id = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{layer_id}{single value specifying an id for the layer}
}
\description{
Clears a kml layer from the map
}

