% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_polygons}
\alias{clear_polygons}
\title{Clear polygon}
\usage{
clear_polygons(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Clears polygons from map
}

