% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_polyline}
\alias{clear_polyline}
\title{Clear polyline}
\usage{
clear_polyline(map)

clear_polyline(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Clears polylines from map

Clears polygons from map
}

