% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_polylines}
\alias{clear_polylines}
\title{Clear polyline}
\usage{
clear_polylines(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Clears polylines from map
}

