% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_search}
\alias{clear_search}
\title{Clear search}
\usage{
clear_search(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
clears the markers placed on the map after using the search box
}
