% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_traffic}
\alias{clear_traffic}
\title{Clear traffic}
\usage{
clear_traffic(map)

clear_traffic(map)
}
\arguments{
\item{map}{a googleway map object}

\item{map}{a googleway map object}
}
\description{
Clears traffic layer from map

Clears transit layer from map
}

