% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{constructURL}
\alias{constructURL}
\title{Construct url}
\usage{
constructURL(map_url, urlArgs)
}
\arguments{
\item{map_url}{string map url}

\item{urlArgs}{other arguments to append to the URL string}
}
\description{
Constructs the relevant API url, given the arguments
}

