% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_datatable.R
\name{DataTableColumn}
\alias{DataTableColumn}
\title{Data Table Column}
\usage{
DataTableColumn(df, id, cols)
}
\arguments{
\item{df}{data.frame}

\item{cols}{columns}
}
\description{
Creates the javascript that's used in a Google Data Table Visualisation from
the specified columns of a data.frame
}

