% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_datatable.R
\name{DataTableHeading}
\alias{DataTableHeading}
\title{Data Table Headings}
\usage{
DataTableHeading(df, cols)
}
\arguments{
\item{df}{data.frame}
}
\description{
Creates the javascript that's required for the headings of a javascript
DataTable
}

