% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_pl.R
\name{decode_pl}
\alias{decode_pl}
\alias{googleway}
\title{decode_pl}
\usage{
decode_pl(encoded)
}
\arguments{
\item{encoded}{String. An encoded polyline}
}
\value{
data.frame of lat/lon coordinates
}
\description{
Decodes a polyline that's generated from google's directions api \url{https://developers.google.com/maps/documentation/directions/}
}
\examples{
## polyline joining the capital cities of Australian states
pl <- "nnseFmpzsZgalNytrXetrG}krKsaif@kivIccvzAvvqfClp~uBlymzA~ocQ}_}iCthxo@srst@"

df_polyline <- decode_pl(pl)
}

