% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodepl.R
\name{decodepl}
\alias{decodepl}
\title{Decodepl}
\usage{
decodepl(encoded)
}
\arguments{
\item{encoded}{String. An encoded polyline}
}
\value{
data.frame of lat/lon coordinates
}
\description{
Decodes a polyline that's generated from google's directions api \url{https://developers.google.com/maps/documentation/directions/}
}
\examples{
## polyline joining the capital cities of Australian states
pl <- "nnseFmpzsZgalNytrXetrG}krKsaif@kivIccvzAvvqfClp~uBlymzA~ocQ}_}iCthxo@srst@"

df_polyline <- decodepl(pl)
}

