% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_instructions}
\alias{direction_instructions}
\title{Direction Instructions}
\usage{
direction_instructions(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Instructions
}
