% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_legs}
\alias{direction_legs}
\title{Direction Legs}
\usage{
direction_legs(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Legs
}
