% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_points}
\alias{direction_points}
\title{Direction Points}
\usage{
direction_points(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Points
}
