% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_polyline}
\alias{direction_polyline}
\title{Direction Polyline}
\usage{
direction_polyline(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Polyline
}
