% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_routes}
\alias{direction_routes}
\title{Direction Routes}
\usage{
direction_routes(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Routes
}
