% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_directions.R
\name{direction_steps}
\alias{direction_steps}
\title{Direction Steps}
\usage{
direction_steps(res)
}
\arguments{
\item{res}{The result of a \code{google_directions} query}
}
\description{
Direction Steps
}
