% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_geojson.R
\name{drag_drop_geojson}
\alias{drag_drop_geojson}
\title{Drag Drop Geojson}
\usage{
drag_drop_geojson(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
A function that enables you to drag GeoJSON and drop it onto a map
}
