% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline.R
\name{encode_pl}
\alias{encode_pl}
\title{Encode PL}
\usage{
encode_pl(lat, lon)
}
\arguments{
\item{lat}{vector of latitude coordinates}

\item{lon}{vector of longitude coordinates}
}
\value{
string encoded polyline
}
\description{
Encodes a series of lat/lon coordinates that specify a path into an encoded polyline
}
\note{
An encoded polyline is generated from google's polyline encoding algorithm (\url{https://developers.google.com/maps/documentation/utilities/polylinealgorithm}).
}
\examples{
encode_pl(lat = c(38.5, 40.7, 43.252), lon = c(-120.2, -120.95, -126.453))
## "_p~iF~ps|U_ulLnnqC_mqNvxq`@"


}
\seealso{
\link{decode_pl}
}
