% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{find_lat_column}
\alias{find_lat_column}
\title{Find Lat Column}
\usage{
find_lat_column(names, calling_function, stopOnFailure = TRUE)
}
\arguments{
\item{names}{string of column names}

\item{stopOnFailure}{logical}

\item{calling_funciton}{the function that called this function}
}
\description{
Tries to identify the latitude column
}

