% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{find_lon_column}
\alias{find_lon_column}
\title{Find Lon Column}
\usage{
find_lon_column(names, calling_function, stopOnFailure = TRUE)
}
\arguments{
\item{names}{string of column names}

\item{calling_function}{the function that called this function}

\item{stopOnFailure}{logical}
}
\description{
Tries to identify the longitude column
}

