% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_melbourne}
\alias{geo_melbourne}
\title{geo_melbourne}
\format{
An object of class \code{json} (inherits from \code{geo_json}) of length 1.
}
\usage{
geo_melbourne
}
\description{
GeoJSON data of Melbourne's Inner suburbs.
}
\details{
This is a subset of the \code{melbourne} data.
}
\keyword{datasets}
