% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_geocode.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode Address}
\usage{
geocode_address(res)
}
\arguments{
\item{res}{The result of a \code{google_geocode} query}
}
\description{
Geocode Address
}
