% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_geocode.R
\name{geocode_coordinates}
\alias{geocode_coordinates}
\title{Geocode Coordinates}
\usage{
geocode_coordinates(res)
}
\arguments{
\item{res}{The result of a \code{google_geocode} query}
}
\description{
Geocode Coordinates
}
