% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route.R
\name{get_route}
\alias{get_route}
\title{GetRoute}
\usage{
get_route(origin, destination, mode = c("driving", "walking", "bicycling",
  "transit"), departure_time = NULL, arrival_time = NULL,
  waypoints = NULL, alternatives = FALSE, avoid = NULL,
  units = c("metric", "imperial"), traffic_model = NULL,
  transit_mode = NULL, transit_routing_preference = NULL, language = NULL,
  region = NULL, key = NULL, output_format = c("data.frame", "JSON"))
}
\description{
This function is deprecated. Please use \link{google_directions}
}

