% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_elevation.R
\name{google_elevation}
\alias{google_elevation}
\title{Google elevation}
\usage{
google_elevation(locations, key, output_format = c("data.frame", "JSON"))
}
\arguments{
\item{key}{string A valid Google Developers Directions API key}

\item{output_format}{string Either 'data.frame' or 'JSON'}
}
\description{
The Google Maps Elevation API provides elevation data for all locations on the surface of the earth, including depth locations on the ocean floor (which return negative values).
}

