% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{google_map}
\alias{google_map}
\title{Google map}
\usage{
google_map(width = NULL, height = NULL, padding = 0)
}
\description{
Generates a google map object. Will only work with a valid key (see \link{map_key})
}
\examples{
\dontrun{

map <- google_map()
map <- map_key(map, key)

## using pipes
library(magrittr)
google_map() \%>\% map_key(key)

}

}

