% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{google_map_directions}
\alias{google_map_directions}
\title{Google Map Directions}
\usage{
google_map_directions(origin = NULL, origin_place_id = NULL,
  destination = NULL, destination_place_id = NULL, travel_mode = NULL,
  dir_action = NULL, waypoints = NULL, waypoint_place_ids = NULL)
}
\arguments{
\item{waypoint_place_ids}{}
}
\description{
Opens a Google Map with the resutls of the specified directions query
}
