% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{google_map_search}
\alias{google_map_search}
\title{Google Map Search}
\usage{
google_map_search(search)
}
\arguments{
\item{search}{string to search for}
}
\description{
Opens a Google Map with the result of the specified search query.
}
\examples{
\dontrun{

google_map_search("Melbourne, Victoria")

google_map_search("Restaruants")

}

}
