% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{google_map_view}
\alias{google_map_view}
\title{google map view}
\usage{
google_map_view(map, location, zoom)
}
\arguments{
\item{map}{a googleway map object created from google_map()}

\item{location}{\code{numeric} vector of latitude/longitude (in that order)
coordinates for the initial starting position of the map.
The map will automatically set the location and zoom if data is added through one
of the various \code{add_} functions. If null, the map will default to Melbourne, Australia.}

\item{zoom}{\code{integer} representing the zoom level of the map (0 is fully zoomed out)}
}
\description{
google map view
}
