% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_roads.R
\name{google_nearestRoads}
\alias{google_nearestRoads}
\title{Nearest Roads}
\usage{
google_nearestRoads(simplify = TRUE, key)
}
\arguments{
\item{simplify}{\code{logical} Inidicates if the returned JSON should be coerced into a list}

\item{key}{\code{string} A valid Google Developers Places API key}
}
\description{
Stuff...
}

