% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_reverse_geocode.R
\name{google_reverse_geocode}
\alias{google_reverse_geocode}
\title{Google reverse geocoding}
\usage{
google_reverse_geocode(location, result_type = NULL, location_type = NULL,
  language = NULL, key, output_format = c("data.frame", "JSON"))
}
\arguments{
\item{location}{numeric Vector of lat/lon coordinates}

\item{result_type}{string vector One or more address types. See \url{https://developers.google.com/maps/documentation/geocoding/intro#Types} for list of available types.}

\item{location_type}{string vector Specifying a location type will restrict the results to this type. If multiple types are specified, the API will return all addresses that match any of the types}

\item{language}{string specifies the language in which to return the results. See the list of supported languages: \url{https://developers.google.com/maps/faq#using-google-maps-apis} If no langauge is supplied, the service will attempt to use the language of the domain from which the request was sent}

\item{key}{string A valid Google Developers Directions API key}

\item{output_format}{string Either 'data.frame' or 'JSON'}
}
\value{
Either data.frame or JSON string of the geocoded address
}
\description{
Geocoding is the process of converting addresses (like "1600 Amphitheatre Parkway, Mountain View, CA") into geographic coordinates (like latitude 37.423021 and longitude -122.083739), which you can use to place markers on a map, or position the map.
}
\examples{
\dontrun{
## searching for the street address for the rooftop location type
google_reverse_geocode(location = c(-37.81659, 144.9841),
                       result_type = c("street_address"),
                       location_type = "rooftop",
                       key = key)
}
}

