% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_timezone.R
\name{google_timezone}
\alias{google_timezone}
\title{Google timezone}
\usage{
google_timezone(location, timestamp = Sys.time(),
  output_format = c("data.frame", "JSON"), key = NULL)
}
\arguments{
\item{location}{vector of lat/lon pair}

\item{timestamp}{POSIXct The Google Maps Time Zone API uses the timestamp to determine whether or not Daylight Savings should be applied. Will default to the current system time}

\item{output_format}{string Either 'data.frame' or 'JSON'}

\item{key}{string A valid Google Developers Directions API key}
}
\description{
The Google Maps Time Zone API provides time offset data for locations on the surface of the earth. You request the time zone information for a specific latitude/longitude pair and date
}

