% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{latitude_column}
\alias{latitude_column}
\title{Latitude column}
\usage{
latitude_column(data, lat, calling_function)
}
\arguments{
\item{data}{\code{data.frame}}

\item{lat}{string identifying the latitude column}

\item{calling_function}{the function that called this function}
}
\description{
calls the correct function to check for latitude column
}

