% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{longitude_column}
\alias{longitude_column}
\title{Longitude column}
\usage{
longitude_column(data, lon, calling_function)
}
\arguments{
\item{data}{\code{data.frame}}

\item{lon}{string identifying the longitude column}

\item{calling_function}{the function that called this function}
}
\description{
calls the correct function to check for longitude column
}

