% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map.R
\name{map_key}
\alias{map_key}
\title{Map key}
\usage{
map_key(map, key)
}
\arguments{
\item{key}{\code{string} a valid Google Maps API key}
}
\description{
Adds the Google Maps API key to a map object
}
\examples{
\dontrun{

map <- google_map()
map$dependencies <- map_key(key)

}
}

