% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{map_styles}
\alias{map_styles}
\title{Map Styles}
\usage{
map_styles()
}
\value{
list of styles
}
\description{
Various style options for maps
}
\note{
you can generate your own map styles at \url{https://mapstyle.withgoogle.com/}
}

