% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{objectColumns}
\alias{objectColumns}
\title{Object Columns}
\usage{
objectColumns(obj = c("polylinePolyline", "polylineCoords", "polygonPolyline",
  "polygonCoords"))
}
\arguments{
\item{obj}{string specifying the type of object}
}
\value{
vector of column names
}
\description{
Defines the columns used by the Maps API so only those required
are kept
}
