% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_polyline.R
\name{objPolygonCoords}
\alias{objPolygonCoords}
\title{Obj Polygon Coords}
\usage{
objPolygonCoords(obj, ids, otherColumns)
}
\arguments{
\item{obj}{\code{data.frame} consisting of columns 'lat' and 'lng'}

\item{ids}{vector of ids for each path/line}

\item{otherColumns}{vector of attribute columns (e.g. stroke_fill, stroke_weight)
that are to be kept in the output}
}
\description{
Creates a list object that can be converted into the required JSON for plotting
coordinates as polygons
}

