% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleway-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{A google map and a layer to add to it}
}
\description{
Uses the pipe operator (\code{\%>\%}) to chain statements. Useful for adding
layers to a \code{google_map}
}
\examples{
\dontrun{

key <- "your_api_key"
google_map(key = key) \%>\%
add_traffic()

}
}

