% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_places.R
\name{place_fields}
\alias{place_fields}
\title{Place Fields}
\usage{
place_fields()
}
\description{
Convenience function to return all the valid basic field values for use in a \link{google_find_place}
search
}
