% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{polygon_transform}
\alias{polygon_transform}
\title{Polygon transform}
\usage{
polygon_transform(SPDF, id_field, CRSObj = "+init=epsg:4326")
}
\arguments{
\item{SPDF}{Spatial Polygons data frame to be encoded}

\item{id_field}{field in @data that identifies the object}

\item{CRSObj}{specify to also change the projection (default output is google map friendly EPSG:4326)}
}
\value{
data.frame with id field, encoded poly lines, `hole` boolean and `ringDir` info
}
\description{
Encodes and foramts a \code{SpatialPolygonDataFrame} object into a Google-friendly \code{data.frame} of encoded polylines
}

