% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_drawing.R
\name{remove_drawing}
\alias{remove_drawing}
\title{Remove drawing}
\usage{
remove_drawing(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
Removes the drawing controls from a map
}
