% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{set_key}
\alias{set_key}
\title{Set Key}
\usage{
set_key(api_key, api = c("api_key", "map_key", "directions_key",
  "distance_key", "elevation_key", "geocode_key", "places_key",
  "places_autocomplete_key", "places_details_key", "roads_key",
  "streetview_key", "timezone_key"))
}
\arguments{
\item{api_key}{Google API key}

\item{api}{The api for which the key applies. If NULL, the \code{api_key} is
assumed to apply to all APIs}
}
\description{
Sets an API key so it's available for all API calls. See details
}
