% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{SetDefault}
\alias{SetDefault}
\title{Set Defaults}
\usage{
SetDefault(col, val, df)
}
\arguments{
\item{col}{column to check / add to the data}

\item{val}{default value for the column}

\item{df}{data to be checked/ added to}
}
\description{
Set Defaults
}

