% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{stripColumns}
\alias{stripColumns}
\title{Strip Columns}
\usage{
stripColumns(data, keep = c("id", "geodesic", "stroke_colour",
  "stroke_weight", "stroke_opacity", "info_window", "mouse_over",
  "mouse_over_group", "layer_id"))
}
\arguments{
\item{data}{data being passed to the browser}

\item{keep}{vector of columns to keep}
}
\description{
Removes unnecessary columns before being passed to browser
}

