% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tram_route}
\alias{tram_route}
\title{Tram Route}
\format{A data frame with 55 observations and 3 variables
\describe{
  \item{shape_pt_lat}{the latitude of each point in the route}
  \item{shape_pt_lon}{the longitude of each point in the route}
  \item{shape_pt_sequence}{the position in the sequence of coordinates for each point}
}}
\usage{
tram_route
}
\description{
The latitude and longitude coordinates specifying the path tram 35 follows in Melbourne.
}
\details{
The data is taken from the PTV GTFS data
}
\keyword{datasets}

