% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tram_stops}
\alias{tram_stops}
\title{Tram stops along tram route 35 in Melbourne}
\format{A data frame with 41 observations and 4 variables
\describe{
  \item{stop_id}{unique ID for each stop}
  \item{stop_name}{the name of each stop}
  \item{stop_lat}{the latitude of the stop}
  \item{stop_lon}{the longitude of the stop}
}}
\usage{
tram_stops
}
\description{
A data set containing the latitude and longitude coordinates of tram stops along route 35 in Melbourne.
}
\details{
The data is taken from the PTV GTFS data
}
\keyword{datasets}
