% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_geojson.R
\name{update_geojson}
\alias{update_geojson}
\title{update geojson}
\usage{
update_geojson(map, layer_id = NULL, style)
}
\arguments{
\item{style}{}
}
\description{
Updates a geojson layer by a specified style
}
\details{
The style object can either be a valid JSON string, or a named list.
The style object will contain the following fields

\itemize{
\item{property : the property of the geoJSON that contains the \code{value}}
\item{value : the value of the geoJSON that identifies the feature to be updated}
\item{features : a list (or JSON object) of features to be updated}
}

see \link{add_geojson} for valid features
}
\examples{
\dontrun{

style <- '{
    "property" : "value",
    "value" : 100,
    "operator" : ">=",
    "features" : {
      "fillColor" : "white",
      "strokeColor" : "white",
      "icon" : "https://developers.google.com/maps/documentation/javascript/examples/full/images/beachflag.png"
    }
  }'

google_map(key = mapKey) \%>\%
    add_geojson(geojson = geojson_txt) \%>\%
    update_geojson(style = lst_style)

lst_style <- list(property = "value", operator = ">=", value = 100,
                  features = list(fillColor = "white",
                                  strokeColor = "white",
                                  icon = "https://developers.google.com/maps/documentation/javascript/examples/full/images/beachflag.png"))

}

google_map(key = mapKey) \%>\%
    add_geojson(geojson = geojson_txt) \%>\%
    update_geojson(style = lst_style)

}
