% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{update_polygons}
\alias{update_polygons}
\title{Update polygons}
\usage{
update_polygons(map, data, id, stroke_colour = NULL, stroke_weight = NULL,
  stroke_opacity = NULL, fill_colour = NULL, fill_opacity = NULL,
  layer_id = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data.frame containing the new values for the polygons}

\item{id}{string representing the column of \code{data} containing the id
values for the polygons. The id values must be present in the data supplied
to \code{add_polygons} in order for the polygons to be udpated}

\item{stroke_colour}{either a string specifying the column of \code{data}
containing the stroke colour of each circle, or a valid hexadecimal numeric
HTML style to be applied to all the circles}

\item{stroke_weight}{either a string specifying the column of \code{data}
containing the stroke weight of each circle, or a number indicating the width of
pixels in the line to be applied to all the circles}

\item{stroke_opacity}{either a string specifying the column of \code{data}
containing the stroke opacity of each circle, or a value between 0 and 1 that
will be aplied to all the circles}

\item{fill_colour}{either a string specifying the column of \code{data}
containing the fill colour of each circle, or a valid hexadecimal numeric
HTML style to be applied to all the cirlces}

\item{fill_opacity}{either a string specifying the column of \code{data}
containing the fill opacity of each circle, or a value between 0 and 1 that
will be aplied to all the circles}

\item{layer_id}{single value specifying an id for the layer.}
}
\description{
Updates specific colours and opacities of specified polygons. Designed to be
used in a shiny application.
}
\note{
Any polygons (as specified by the \code{id} argument) that do not exist
in the \code{data} passed into \code{add_polygons()} will not be added to the map.
This function will only update the polygons that currently exist on the map
when the function is called.
}

