% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{update_style}
\alias{update_style}
\title{Update style}
\usage{
update_style(map, style = NULL)
}
\arguments{
\item{map}{googleway map object}

\item{style}{JSON specifying the style features to apply to the map.}
}
\description{
Updates the map with the given style
}

