% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_circles}
\alias{add_circles}
\title{Add circle}
\usage{
add_circles(map, data, lat = NULL, lon = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame}
}
\description{
Add circles to a google map
}

