% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_dragdrop}
\alias{add_dragdrop}
\title{Drag Drop Geojson}
\usage{
add_dragdrop(map)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}
}
\description{
A function that enables you to drag data and drop it onto a map. Currently
only supports GeoJSON files / text
}
