% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_fusion}
\alias{add_fusion}
\title{Add Fusion}
\usage{
add_fusion(map, fusion, table_id, layer_id = NULL)
}
\arguments{
\item{table_id}{Encrypted Fusion Table ID}

\item{layer_id}{}
}
\description{
Adds a fusion table to a map
}
