% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_markers}
\alias{add_markers}
\title{Add markers}
\usage{
add_markers(map, data, lat = NULL, lon = NULL, title = NULL,
  draggable = NULL, opacity = NULL, label = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude.}

\item{lat}{string specifying the column of \code{data} containing the 'latitude' coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude' coordinates. If left NULL, a best-guess will be made}

\item{title}{string specifying the column of \code{data} containing the 'title' of the markers}

\item{draggable}{string specifying the column of \code{data} defining if the marker is 'draggable' (either TRUE or FALSE)}

\item{opacity}{string specifying the column of \code{data} defining the 'opacity' of the maker. Values must be between 0 and 1 (inclusive)}

\item{label}{string specifying the column of \code{data} defining the character to appear in the centre of the marker. Values will be coerced to strings, and only the first character will be used.}
}
\description{
Add markers to a google map
}

