% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polygon}
\alias{add_polygon}
\title{Add polygon}
\usage{
add_polygon(map, data, lineSource = c("coords", "polyline"), group = NULL,
  group_options = NULL, lat = NULL, lon = NULL, polyline = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{lineSource}{string specifying the type of source data for the line, one of either 'coords' (for coordinates) or 'polyline'}

\item{group}{... placeholder...}

\item{group_options}{... placeholder...}

\item{lat}{string specifying the column of \code{data} containing the 'latitude' coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude' coordinates. If left NULL, a best-guess will be made}
}
\description{
Add a polygon to a google map
}

