% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polygons}
\alias{add_polygons}
\title{Add polygon}
\usage{
add_polygons(map, data = get_map_data(map), polyline, stroke_colour = NULL,
  stroke_weight = NULL, stroke_opacity = NULL, fill_colour = NULL,
  fill_opacity = NULL, info_window = NULL, mouse_over = NULL,
  mouse_over_group = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{polyline}{string specifying the column containing the polyline}

\item{stroke_colour}{either a string specifying the column of \code{data} containing the stroke colour of each circle, or a valid hexadecimal numeric HTML style to be applied to all the circles}

\item{stroke_weight}{either a string specifying the column of \code{data} containing the stroke weight of each circle, or a number indicating the width of pixels in the line to be applied to all the circles}

\item{stroke_opacity}{either a string specifying the column of \code{data} containing the stroke opacity of each circle, or a value between 0 and 1 that will be aplied to all the circles}

\item{fill_colour}{either a string specifying the column of \code{data} containing the fill colour of each circle, or a valid hexadecimal numeric HTML style to be applied to all the cirlces}

\item{fill_opacity}{either a string specifying the column of \code{data} containing the fill opacity of each circle, or a value between 0 and 1 that will be aplied to all the circles}

\item{info_window}{string specifying the column of data to display in an info window when a polygon is clicked}

\item{mouse_over}{string specifying the column of data to display when the mouse rolls over the polygon}

\item{mouse_over_group}{string specifying the column of data specifying which groups of polygons to highlight on mouseover}
}
\description{
Add a polygon to a google map.
}
\examples{
\dontrun{

## polygon with a hole - Bermuda triangle
pl_outer <- gepaf::encodePolyline(data.frame(lat = c(25.774, 18.466,32.321),
                                            lng = c(-80.190, -66.118, -64.757)))

pl_inner <- gepaf::encodePolyline(data.frame(lat = c(28.745, 29.570, 27.339),
                                             lng = c(-70.579, -67.514, -66.668)))


l <- list(c(pl_outer, pl_inner))

df <- data.frame(polyline = rep(NA, length(l)))

df$polyline <- l

google_map(key = map_key, height = 800, location = c(25.774, -80.190), zoom = 3) \%>\%
  add_polygons(data = df, polyline = "polyline", mouse_over = "polyline")


}

}

