% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polyline}
\alias{add_polyline}
\title{Add polyline}
\usage{
add_polyline(map, data)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude.}
}
\description{
Add a polyline to a google map
}

