% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polyline}
\alias{add_polyline}
\title{Add polyline}
\usage{
add_polyline(map, data, group = NULL, group_options = NULL, lat = NULL,
  lon = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{group}{... placeholder...}

\item{group_options}{... placeholder ... colours etc}

\item{lat}{string specifying the column of \code{data} containing the 'latitude' coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude' coordinates. If left NULL, a best-guess will be made}
}
\description{
Add a polyline to a google map
}

