% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{add_polylines}
\alias{add_polylines}
\title{Add polyline}
\usage{
add_polylines(map, data = get_map_data(map), polyline, geodesic = NULL,
  stroke_colour = NULL, stroke_weight = NULL, stroke_opacity = NULL,
  info_window = NULL, mouse_over = NULL, mouse_over_group = NULL,
  update_map_view = TRUE, layer_id = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the latitude coordinates, and the other specifying the longitude. If Null, the data passed into \code{google_map()} will be used.}

\item{polyline}{string specifying the column of \code{data} containing the encoded 'polyline'}

\item{geodesic}{logical}

\item{stroke_colour}{either a string specifying the column of \code{data} containing the stroke colour of each circle, or a valid hexadecimal numeric HTML style to be applied to all the circles}

\item{stroke_weight}{either a string specifying the column of \code{data} containing the stroke weight of each circle, or a number indicating the width of pixels in the line to be applied to all the circles}

\item{stroke_opacity}{either a string specifying the column of \code{data} containing the stroke opacity of each circle, or a value between 0 and 1 that will be aplied to all the circles}

\item{info_window}{string specifying the column of data to display in an info window when a polygon is clicked}

\item{mouse_over}{string specifying the column of data to display when the mouse rolls over the polygon}

\item{mouse_over_group}{string specifying the column of data specifying which groups of polylines to highlight on mouseover}

\item{update_map_view}{logical specifying if the map should re-centre according to the polyline.}

\item{layer_id}{single value specifying an id for the layer.}
}
\description{
Add a polyline to a google map
}
\note{
using \code{update_map_view = TRUE} for multiple polylines may be slow, so it may be more appropriate to set the view of the map using the location argument of \code{google_map()}
}
\examples{
\dontrun{

library(magrittr)
pl <- "pizeF_{~sZjLhAdAJLwBJkBv@wN~Bgd@x@sNjAeUR_FTkEAqBGiAQuAw@{E_C}NGwA?sANaELkDDu@LgAfAiF?g@EI?EIo@EMMKe@UQGaAM}IaA_Fg@_JaAsEc@_"
df <- data.frame(polyline = pl)
google_map(key = map_key) \%>\%
  add_polylines(data = df, polyline = "polyline")


}
}

