% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear}
\alias{clear}
\alias{clear_bicycling}
\alias{clear_circles}
\alias{clear_heatmap}
\alias{clear_markers}
\alias{clear_polygons}
\alias{clear_polylines}
\alias{clear_rectangles}
\alias{clear_traffic}
\alias{clear_transit}
\title{clear map elements}
\usage{
clear_markers(map, layer_id = NULL)

clear_circles(map, layer_id = NULL)

clear_heatmap(map, layer_id = NULL)

clear_traffic(map)

clear_transit(map)

clear_bicycling(map)

clear_polylines(map, layer_id = NULL)

clear_polygons(map, layer_id = NULL)

clear_rectangles(map, layer_id = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{layer_id}{id value of the layer to be removed from the map}
}
\description{
clears elements from a map
}
\note{
These operations are intended for use in conjunction with
\link{google_map_update} in an interactive shiny environment
}

