% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{clear_bicycling}
\alias{clear_bicycling}
\title{Clear bicycling}
\usage{
clear_bicycling(map)
}
\arguments{
\item{map}{a googleway map object}
}
\description{
Clears bicycle route layer from map
}

