% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layers.R
\name{find_lat_column}
\alias{find_lat_column}
\title{This is taken directly from Rstudio/leaflet. I would like to make use of some of the functions}
\usage{
find_lat_column(names, calling_function, stopOnFailure = TRUE)
}

