% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route.R
\name{get_route}
\alias{get_route}
\title{GetRoute}
\usage{
get_route()
}
\value{
Either data.frame or JSON string of the route between origin and destination
}
\description{
This function is deprecated. Please use \link{google_directions}
}

