% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route.R
\name{get_route}
\alias{get_route}
\title{GetRoute}
\usage{
get_route(...)
}
\arguments{
\item{...}{arguments passed to google_directions. Legacy - so that old 'get_route() calls will still enter this function.}
}
\description{
This function is deprecated. Please use \link{google_directions}
}
