% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{google_charts}
\alias{google_charts}
\title{Google Charts}
\description{
Google Charts that can be displayed
}
\section{Pie Chart}{


A Pie chart requires a \code{data.frame} of two columns. The first column contains
the text that's used as the label, and the second contains the data used in
the pie.

type: 'pie'

options
\itemize{
  \item{title - string of text}
  \item{is3D - logical}
}
}

\section{Area}{


\url{https://developers.google.com/chart/interactive/docs/gallery/areachart}

Each row of data represents a data point at the same x-axis location
}

