% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{google_map_update}
\alias{google_map_update}
\title{Google map update}
\usage{
google_map_update(map_id, session = shiny::getDefaultReactiveDomain(),
  data = NULL, deferUntilFlush = TRUE)
}
\arguments{
\item{map_id}{string The output ID of the map in a shiny application}

\item{session}{the Shiny session object to which the map belongs; usually the default value will suffice}
}
\description{
Update a map in a shiny app. Use this function whenever the map needs to respond to reactive content.
}

