% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_utilities.R
\name{google_map_update}
\alias{google_map_update}
\title{Google map update}
\usage{
google_map_update(map_id, session = shiny::getDefaultReactiveDomain(),
  data = NULL, deferUntilFlush = TRUE)
}
\description{
update a map in a shiny app
}

