% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_map_layer_heatmap.R
\name{update_heatmap}
\alias{update_heatmap}
\title{update heatmap}
\usage{
update_heatmap(map, data, lat = NULL, lon = NULL, weight = NULL,
  option_gradient = NULL, layer_id = NULL, digits = 4, legend = F,
  legend_options = NULL)
}
\arguments{
\item{map}{a googleway map object created from \code{google_map()}}

\item{data}{data frame containing at least two columns, one specifying the
latitude coordinates, and the other specifying the longitude. If Null,
the data passed into \code{google_map()} will be used.}

\item{lat}{string specifying the column of \code{data} containing the 'latitude'
coordinates. If left NULL, a best-guess will be made}

\item{lon}{string specifying the column of \code{data} containing the 'longitude'
coordinates. If left NULL, a best-guess will be made}

\item{weight}{string specifying the column of \code{data} containing the 'weight'
associated with each point. If NULL, each point will get a weight of 1.}

\item{option_gradient}{vector of colours to use as the gradient colours in the legend.
See details}

\item{layer_id}{single value specifying an id for the layer.}

\item{digits}{integer. Use this parameter to specify how many digits (decimal places)
should be used for the latitude / longitude coordinates.}
}
\description{
updates a heatmap layer
}
\details{
The \code{option_gradient} is only used to craete the legend, and not to change
the colours of the heat layer. If you are not displaying
a legend this argument is not needed. If you are displaying a legend, you should
provide the same gardient as in the \code{add_heatmap} call.
}
\examples{
\dontrun{

map_key <- 'your_api_key'

set.seed(20170417)
df <- tram_route
df$weight <- sample(1:10, size = nrow(df), replace = T)

google_map(key = map_key, data = df) \%>\%
 add_heatmap(lat = "shape_pt_lat", lon = "shape_pt_lon", weight = "weight",
              option_radius = 0.001)

## update by adding the same data again to double the number of points at each location
df_update <- df
google_map(key = map_key, data = df) \%>\%
 add_heatmap(lat = "shape_pt_lat", lon = "shape_pt_lon", weight = "weight",
              option_radius = 0.001) \%>\%
 update_heatmap(df_update, lat = "shape_pt_lat", lon = "shape_pt_lon")

}

}
